+++ src/backend/github/workflow/workflow.test.ts
@@ -10,7 +10,7 @@ describe("GitHub Actions", () => {
 			jest.setTimeout(20000);
 		});
 		it("should update the Jira issue with the linked GitHub workflow_run", async () => {
-			const fixture = require("../../../common/test-utils/fixtures/workflow-basic.json");
+			const fixture = require("../../../../test-utils/fixtures/workflow-basic.json");
 
 			githubNock
 				.get("/users/test-workflow-run-user-login")
