+++ src/backend/sync/installation/installation.test.ts
@@ -2,8 +2,8 @@
 import { sortedRepos } from ".";
 
 describe("Sync helpers suite", () => {
-	const repoSyncState = require("../../../common/test-utils/fixtures/repo-sync-state.json");
-	const sortedReposFunc = require("../../../common/test-utils/fixtures/sorted-repos.json");
+	const repoSyncState = require("../../../../test-utils/fixtures/repo-sync-state.json");
+	const sortedReposFunc = require("../../../../test-utils/fixtures/sorted-repos.json");
 
 	it("sortedRepos should sort repos by updated_at", () => {
 		expect(sortedRepos(repoSyncState)).toEqual(sortedReposFunc);
